requirements.txt
asgiref==3.5.2
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
ecdsa==0.18.0
idna==3.3
Markdown==3.4.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
